update views.py update test_reports.py update main.py update test_views.py create utils.py create test_utils.py
